user-service/src/main/java/com/kanwise/user_service/validation/logic/common/NullOrNotBlankValidator.java
package com.kanwise.user_service.validation.logic.common;


import com.kanwise.user_service.validation.annotation.common.NullOrNotBlank;
import org.springframework.stereotype.Service;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;

@Service
public class NullOrNotBlankValidator implements ConstraintValidator<NullOrNotBlank, String> {

    @Override
    public boolean isValid(String value, ConstraintValidatorContext constraintValidatorContext) {
        return value == null || value.trim().length() > 0;
    }
}
